refactor(import): apply review findings on charset handling and layout
- fetch_url_text honors the server-declared Content-Type charset
  (after BOM, before statistics) instead of dropping the parameter:
  short pages with a small Cyrillic fraction no longer risk silent
  mojibake with no correction path; decode_for_content_type covers it
  with tests incl. BOM-beats-declaration precedence
- import command cluster (picker/reader/fetcher + tests) moves to
  commands/import.rs along its domain seam; mod.rs sheds ~400 lines
- pick_import_file filter now reads SUPPORTED_EXTENSIONS instead of a
  second hardcoded extension list
- dropHandlerRef assignment moved into an effect (committed-render
  discipline); IconChevronDown moved to the shared icons module
- SPA heuristic: boundary test pins extracted-text = threshold - 1
